add stress gauge chart back to dashboard, begin implementing data context in stress widget
